refactor(code structure): TopDownParser
Split `TopDownParser` into separate files for better modularity and
maintainability. Added implementations for expressions, declarations,
and statements handling.
